+++ src/nonlocal.jl
@@ -34,7 +34,7 @@ function _local_bound_single(initial_score::T, chunk, sizeG, offset, squareG::Ar
     oa, ob, ia, ib = sizeG
     score = initial_score
     ind = digits(chunk[1]; base = ob, pad = ib)
-    offset_ind = zeros(Int64, ib)
+    offset_ind = zeros(Int, ib)
     Galice = zeros(T, oa * ia, 1)
     maxvec = zeros(T, 1, ia)
     for b in chunk[1]:chunk[2]
@@ -345,7 +345,7 @@ function tensor_correlation(p::AbstractArray{T, N2}, behaviour::Bool = false; ma
         if all(x.I .> marg)
             FC[x] = sum((-1)^sum(a[n] - 1 for n in 1:N if x[n] > marg; init = 0) * p[a, (x.I .- marg)...] for a in cia)
         else
-            x_colon = Union{Colon, Int64}[x[n] > marg ? x[n] - marg : Colon() for n in 1:N]
+            x_colon = Union{Colon, Int}[x[n] > marg ? x[n] - marg : Colon() for n in 1:N]
             FC[x] = sum((-1)^sum(a[n] - 1 for n in 1:N if x[n] > marg; init = 0) * sum(p[a, x_colon...]) for a in cia)
         end
     end
@@ -353,7 +353,7 @@ function tensor_correlation(p::AbstractArray{T, N2}, behaviour::Bool = false; ma
         FC ./= 2^N
     elseif marg
         for n in 1:N
-            x_colon = Union{Colon, Int64}[i == n ? 1 : Colon() for i in 1:N]
+            x_colon = Union{Colon, Int}[i == n ? 1 : Colon() for i in 1:N]
             FC[x_colon...] ./= m[n]
         end
     end
